client.py/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["stendler <git@stendler.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mastodon-py = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
